pyproject.toml
[project]
name = "pitchs-edge"
version = "0.1.0"
description = "The Pitch's Edge — soccer betting edge detection"
authors = [{ name = "Wegs Benedict" }]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pandas>=2.2",
    "numpy>=1.26",
    "scipy>=1.12",
    "python-dotenv>=1.0",
    "streamlit>=1.32",
    "altair>=5.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pitchs_edge"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "C4", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
